# openapi/openapi.yaml
...
openapi: 3.1.0
...
components:
  schemas:
...
    Instant:
      type: string
      format: date-time
      examples:
      - 2022-03-10T16:15:50Z
...
    NamedRef:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
...
    PagedSnapshotHistoryResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/SnapshotHistoryResponse"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
...
    SnapshotHistoryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        pipelineId:
          type: integer
          format: int64
        pipelineName:
          type: string
        type:
          $ref: "#/components/schemas/SnapshotType"
        outcome:
          $ref: "#/components/schemas/SnapshotState"
        totalTables:
          type: integer
          format: int32
        completedTables:
          type: integer
          format: int32
        totalRowsScanned:
          type: integer
          format: int64
        startedAt:
          $ref: "#/components/schemas/Instant"
        completedAt:
          $ref: "#/components/schemas/Instant"
        durationSeconds:
          type: integer
          format: int64
        tables:
          type: array
          items:
            $ref: "#/components/schemas/TableHistoryEntry"
    SnapshotState:
      type: string
      enum:
      - IDLE
      - RUNNING
      - PAUSED
      - COMPLETED
      - ABORTED
      - SKIPPED
      - UNKNOWN
    SnapshotType:
      type: string
      enum:
      - INITIAL
      - BLOCKING
      - INCREMENTAL
    SourceRequest:
      type: object
      required:
      - name
      - type
      - schema
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        type:
          type: string
          minLength: 1
        schema:
          type: string
          minLength: 1
        connection:
          $ref: "#/components/schemas/NamedRef"
        vaults:
          type: array
          uniqueItems: true
          items:
            $ref: "#/components/schemas/NamedRef"
        config:
          type: object
          additionalProperties: {}
    SourceResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        type:
          type: string
        schema:
          type: string
        connection:
          $ref: "#/components/schemas/NamedRef"
        vaults:
          type: array
          uniqueItems: true
          items:
            $ref: "#/components/schemas/NamedRef"
        config:
          type: object
          additionalProperties: {}
    TableHistoryEntry:
      type: object
      properties:
        tableName:
          type: string
        outcome:
          $ref: "#/components/schemas/TableState"
        rowsScanned:
          type: integer
          format: int64
        skipReason:
          type: string
        durationSeconds:
          type: integer
          format: int64
    TableState:
      type: string
      enum:
      - PENDING
      - IN_PROGRESS
      - COMPLETED
      - SKIPPED
      - FAILED
...
  securitySchemes:
    SecurityScheme:
      type: openIdConnect
      openIdConnectUrl: https://keycloak.invalid/realms/none/.well-known/openid-configuration
      description: Authentication
paths:
...
  /api/pipelines/{pipelineId}/snapshots/history:
    get:
      summary: List completed snapshot runs for a pipeline (paginated)
      tags:
      - snapshots
      parameters:
      - name: pipelineId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: from
        in: query
        schema:
          $ref: "#/components/schemas/Instant"
      - name: outcome
        in: query
        schema:
          $ref: "#/components/schemas/SnapshotState"
      - name: page
        in: query
        schema:
          type: integer
          format: int32
          minimum: 0
          default: 0
      - name: size
        in: query
        schema:
          type: integer
          format: int32
          maximum: 100
          minimum: 1
          default: 20
      - name: to
        in: query
        schema:
          $ref: "#/components/schemas/Instant"
      - name: type
        in: query
        schema:
          $ref: "#/components/schemas/SnapshotType"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedSnapshotHistoryResponse"
  /api/pipelines/{pipelineId}/snapshots/history/{historyId}:
    get:
      summary: Return a single snapshot run with its per-table breakdown
      tags:
      - snapshots
      parameters:
      - name: historyId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: pipelineId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SnapshotHistoryResponse"
        "404":
          description: Unknown snapshot run for this pipeline
...
  /api/sources/{id}:
    put:
      summary: Updates an existing source
      tags:
      - sources
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SourceRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceResponse"
        "400":
          description: Bad Request
    get:
      summary: Returns a source with given id
      tags:
      - sources
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceResponse"
    delete:
      summary: Deletes an existing source
      tags:
      - sources
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204": {}
...
info:
  title: Debezium Platform API
  version: 3.7.0.Final
  description: REST API for the Debezium Platform
  contact:
    name: Debezium
    url: https://github.com/debezium/debezium-platform
servers:
- url: http://localhost:8080
  description: Auto generated value
- url: http://0.0.0.0:8080
  description: Auto generated value